DXP-1568 Move code for selecting product ids to correct place
